+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Builtin Apps:
 Frameworks:
 
 OS:
+- LilyGo T-Watch S3 Plus: enable audio output